+++ packages/cronjob-rss-feed/rss.js
@@ -1,8 +1,8 @@
+import { config } from './configs.js'
+import { log } from './utils.js'
 import axios from 'axios'
 import RSS from 'rss'
 import { Storage } from '@google-cloud/storage'
-import { IncomingWebhook } from '@slack/webhook'
-import { config } from './configs.js'
 
 const storage =
   config.gcs.projectId && config.gcs.keyFilename
@@ -11,7 +11,6 @@ const storage =
         keyFilename: config.gcs.keyFilename,
       })
     : new Storage()
-const webhook = new IncomingWebhook(config.slackWebhook)
 
 const fetchPosts = async () => {
   try {
@@ -84,73 +83,6 @@ const uploadToGCS = async (data) => {
     )
   } catch (error) {
     await log('Error uploading RSS to GCS', error)
-  }
-}
-
-const sendSlackNotification = async (message, type = 'log') => {
-  try {
-    await webhook.send({
-      blocks: [
-        {
-          type: 'header',
-          text: {
-            type: 'plain_text',
-            emoji: true,
-            text:
-              type === 'error'
-                ? ':warning:  KidsReporter Cronjob'
-                : ':information_source:  KidsReporter Cronjob',
-          },
-        },
-        {
-          type: 'context',
-          elements: [
-            {
-              text: `${type.toUpperCase()}  |  *${new Date().toISOString()}*`,
-              type: 'mrkdwn',
-            },
-          ],
-        },
-        {
-          type: 'divider',
-        },
-        {
-          type: 'section',
-          text: {
-            type: 'mrkdwn',
-            text: message,
-          },
-        },
-        {
-          type: 'divider',
-        },
-        {
-          type: 'context',
-          elements: [
-            {
-              type: 'mrkdwn',
-              text: 'Sent from kids-reporter/kids-reporter-monorepo/packages/cronjob-rss-feed',
-            },
-          ],
-        },
-      ],
-    })
-    console.log('Slack notification sent')
-  } catch (error) {
-    await log('Error sending Slack notification', error)
-  }
-}
-
-const log = async (message, error = '', type = '') => {
-  if (type === '') {
-    type = error === '' ? 'log' : 'error'
-  }
-  if (type === 'error') {
-    console.error(message, error)
-    await sendSlackNotification(`${message}: ${error.toString()}`, type)
-  } else if (type === 'log') {
-    console.log(message, error)
-    await sendSlackNotification(message, type)
   }
 }
